requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pandas>=2.0.0
numpy>=1.24.0
pillow>=10.0.0
pillow-heif>=0.16.0
streamlit>=1.28.0
altair>=5.0.0
pyrender>=0.1.45
trimesh>=4.0.0
captum>=0.7.0
tqdm>=4.0.0h5py>=3.10.0
